+++ CHANGELOG.md
@@ -13,9 +13,12 @@ See [`UPGRADE.md`](./UPGRADE.md) for additional help when upgrading to newer ver
 
 - `NifException` for using Elixir exception structs
 
+## [0.22.1] - 2021-10-05
+
 ### Fixed
 
 - Fix codegen problem for untagged enums (#370)
+- Fix handling local dependencies with `@external_resources` (#381)
 
 ## [0.22.0] - 2021-06-22